Import NESP2_visualization (copy-paste latest version)
The repository from which we would have liked to import the history had large
files in its past, making its size unnecessarily large (>200Mb). An attempt to
remove the corresponding commits and to keep the rest failed. Therefore the
latest version version is drawn from:

https://github.com/catcad/NESP2_visualization

and simply copy-pasted in here, disregarding the history ¯\_(ツ)_/¯. The date
of copy-pasting is 2019-10-28.