Update the selectedStateInfoBoxContent variable
When a new state is selected this variable gets updated.
That way, when the mouse hover out of an unselected state,
the info of the selected state are displayed by default.
This means when the mouse is in the selected state (which is
not part of the geoJSON features layer) or outside Nigeria.