Update cluster toggle rules between national and state level
When going from the national to the state level for the first time, if
the database mention that there is a remotely mapping surveys for the
selected state, then the RMS toggle will be switch to true. If the user
played with the ISSI toggle on the national level, then the option of
the user will be kept.